Export the logic to the Blur component
