+++ src/components/transactions/item.js
@@ -9,40 +9,17 @@ import { B, Small } from '../toolBox/typography';
 import { stringShortener } from '../../utilities/helpers';
 import loadingAnimation from '../../assets/animations/loading-dots.json';
 import transactions from '../../constants/transactions';
+import Blur from './blur';
 import withTheme from '../withTheme';
 import getStyles from './styles';
-import darkMediumOutgoing from '../../assets/images/amountBlur/outgoing/darkMedium.png';
-import darkSmallOutgoing from '../../assets/images/amountBlur/outgoing/darkSmall.png';
-import darkMediumIncoming from '../../assets/images/amountBlur/incoming/darkMedium.png';
-import darkSmallIncoming from '../../assets/images/amountBlur/incoming/darkSmall.png';
-import lightMediumOutgoing from '../../assets/images/amountBlur/outgoing/lightMedium.png';
-import lightSmallOutgoing from '../../assets/images/amountBlur/outgoing/lightSmall.png';
-import lightMediumIncoming from '../../assets/images/amountBlur/incoming/lightMedium.png';
-import lightSmallIncoming from '../../assets/images/amountBlur/incoming/lightSmall.png';
-
-const blurs = {
-  outgoing: {
-    darkMedium: darkMediumOutgoing,
-    darkSmall: darkSmallOutgoing,
-    lightMedium: lightMediumOutgoing,
-    lightSmall: lightSmallOutgoing,
-  },
-  incoming: {
-    darkMedium: darkMediumIncoming,
-    darkSmall: darkSmallIncoming,
-    lightMedium: lightMediumIncoming,
-    lightSmall: lightSmallIncoming,
-  },
-};
-
 
 const txTypes = ['accountInitialization', 'setSecondPassphrase', 'registerDelegate', 'vote'];
 
 class Item extends React.Component {
-  showDetail(tx, account) {
+  showDetail = (tx, account, incognito) => {
     this.props.navigate({
       routeName: 'TxDetail',
-      params: { tx, account },
+      params: { tx, account, incognito },
     });
   }
 
@@ -54,7 +31,7 @@ class Item extends React.Component {
 
   render() {
     const {
-      styles, tx, account, incognito, theme,
+      styles, tx, account, incognito,
     } = this.props;
     let direction = 'incoming';
     let address = tx.senderId;
@@ -65,11 +42,10 @@ class Item extends React.Component {
       addressShortened = stringShortener(tx.recipientId, 10, 3);
     }
     const amount = direction === 'incoming' ? fromRawLsk(tx.amount) : `-${fromRawLsk(tx.amount)}`;
-    const amountSize = amount.length > 2 ? 'Medium' : 'Small';
 
     return (<TouchableOpacity
       style={[styles.itemContainer, styles.theme.itemContainer]}
-      onPress={this.showDetail.bind(this, tx, account)}>
+      onPress={() => this.showDetail(tx, account, incognito)}>
       <View style={styles.innerContainer}>
       <View style={[styles.itemColumn, styles.avatarContainer]}>
         {
@@ -104,10 +80,7 @@ class Item extends React.Component {
         }
         {
           (tx.type === 0 && (tx.recipientId !== tx.senderId)) && incognito ?
-            <View style={styles.amount}>
-              <Image style={styles[`blur${amountSize}`]}
-                source={blurs[direction][`${theme}${amountSize}`]} />
-            </View> : null
+            <Blur value={amount} direction={direction} /> : null
         }
         {
           typeof this.props.tx.timestamp !== 'number' ?
